Split main component into smaller ones
